adding enum roletype

+++ BIS.Common/Entities/BaseEntity.cs
@@ -9,8 +9,8 @@ namespace BIS.Common.Entities
     public class BaseEntity
     {
         public int Id { get; set; }
-        public short CreatedBy { get; set; }
-        public short UpdatedBy { get; set; }
+        public int CreatedBy { get; set; }
+        public int UpdatedBy { get; set; }
         public DateTime CreatedOn { get; set; }
         public DateTime UpdatedOn { get; set; }
         public bool IsActive { get; set; }
